ndfc_common changes for MSD fabrics, and example MSD playbook
For MSD deployments, networks and VRFs need to be added to the MSD fabric, rather than the individual site/child fabrics.  For now, we've added duplicate entries for networks and VRFs for the MSD fabric.  These mirror the entries for the site/child fabrics and will be used if the fabric_name in a playbook corresponds to the MSD fabric.

We may create a separate ndfc_common_msd role in the future but, for now, this was the easier route to take.

We've also added an example playbook to create both site fabrics (f1 and f2) and the MSD fabric.