ndfc_device_merged: Move REST config-save/config-deploy calls into worker.yml
Moving REST calls (which must be called after every device merge) into the worker for ndfc_device_merged.  This allows us to skip these calls if the device already exists (which is likely the most common case).  If the user wants to add the ndfc_rest_config_deploy_all Role back into the example playbooks, it will still work there and be called unconditionally.

TODO: We can add a toggle variable and set it to True by default in defaults.yml.  We could then add this to worker.yml to enable the user to force config-save and config-deploy calls after every device merge.  Will look into this for the next commit.