Change vrf_id value due to a bug in NDFC Default_VRF_Universal template
Default_VRF_Universal template limits the community portion of route-target to to range 0-65535 (Cisco bug ID CSCwc60385).

vrf_id in this repo were set to 9003031 and 9003032 for vrfs v1 and v2 respectively.  These are valid values, but do not work, given CSCwc60385.  We will update these back to their original values when CSCwc60385 is fixed.